fixed the values
