+++ src/ux/UserTest/utils/nasaTlxData.js
@@ -23,10 +23,11 @@ export function getNASATLXData(nasaTlxData) {
   const mostStressful = dimensionEntries.reduce((max, current) => current[1] > max[1] ? current : max)
   const leastStressful = dimensionEntries.reduce((min, current) => current[1] < min[1] ? current : min)
 
-  const overallScores = nasaTlxData.map(r => r.overallScore)
+  // Calculate overall score as average of all 6 dimensions
+  const overallScore = Math.round(((dimensionAverages.mentalDemand + dimensionAverages.physicalDemand + dimensionAverages.temporalDemand + dimensionAverages.performance + dimensionAverages.effort + dimensionAverages.frustration) / 6) * 10) / 10
 
   return {
-    averageOverallScore: Math.round((overallScores.reduce((sum, score) => sum + score, 0) / overallScores.length) * 10) / 10,
+    averageOverallScore: overallScore,
     totalRespondents: nasaTlxData.length,
     mostStressfulDimension: formatDimensionName(mostStressful[0]),
     leastStressfulDimension: formatDimensionName(leastStressful[0]),
